Changed WideFloat.multiplicativeInverse2 to use Float as its "cheat" for getting the inverse's exponent.

+++ Sources/BigMath/WideFloat/WideFloat+FloatingPoint.swift
@@ -466,7 +466,7 @@ extension WideFloat: FloatingPoint
         
         let result = Self(
             significandBitPattern: q.low,
-            exponent: (1 / self.significand.doubleValue).exponent
+            exponent: (1 / self.significand.floatValue).exponent
         )
         assert(!result.isNegative)
         assert(result.isNormalized)
